renamed component folder
